+++ src/Resources/config/services/controller.xml
@@ -67,8 +67,7 @@
         </service>
 
         <service id="bitbag_sylius_mollie_plugin.controller.action.admin.change_position_payment_method_action" class="BitBag\SyliusMolliePlugin\Controller\Action\Admin\ChangePositionPaymentMethodAction">
-            <argument type="service" id="bitbag_sylius_mollie_plugin.repository.mollie_gateway_config"/>
-            <argument type="service" id="doctrine.orm.entity_manager"/>
+            <argument type="service" id="bitbag_sylius_mollie_plugin.creator.change_position_payment_method_creator"/>
         </service>
     </services>
 </container>
